issue-45 - upgrade dependencies. Add .idea and issue-sync (output of go build) to .gitignore. Specify int64 where necessary.

+++ lib/issues.go
@@ -32,7 +32,7 @@ func CompareIssues(config cfg.Config, ghClient clients.GitHubClient, jiraClient 
 		return nil
 	}
 
-	ids := make([]int, len(ghIssues))
+	ids := make([]int64, len(ghIssues))
 	for i, v := range ghIssues {
 		ids[i] = v.GetID()
 	}
